distributed_agent_platform/requirements.txt
# Core Framework
pyautogen==0.2.12
openai>=1.6.1,<2.0.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# Message Queue & Caching
redis==5.0.1
hiredis==2.2.3

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

